Add support for FreeBSD.
This implementation requires to mount the Linux compatibility /proc
filesystem in /compat/linux/proc. Check the FreeBSD handbook for
details.
